requirements.txt
selenium
bs4
requests
python-dotenv
pandas